python/requirements.txt
fastapi
uvicorn
spacy
python-multipart
python-docx
PyPDF2
rake-nltk
python-dotenv
openai
transformers
torch
sentencepiece
textblob
sentence-splitter
protobuf
